compose: add ostree commit image types
These are a bit different than our other image types, but this is the
closest fit I could find without majorly reworking the API/UI.

The image types are not yet part of osbuild-composer, but I think we can
merge this eagerly as they will be ignored in case the backend doesn't
support them.

Two new image types are introduced "Fedora IoT" and "RHEL for Edge",
these are going to be very similar, but obviously available for
Fedora/RHEL respectively.

Signed-off-by: Tom Gundersen <teg@jklm.no>