Updates how warnings are handled
If password one has a value and the user is shifting focus
to password two, we don't display a warning. But if the user
leave password two empty, we need to display a warning at
that time.

The warning should also display in the case where password
two has a value, but password one does not.